backtest_gui/requirements.txt
PyQt5>=5.15.0
matplotlib>=3.4.0
pandas>=1.2.0
numpy>=1.20.0
psycopg2-binary>=2.8.6
PyYAML>=6.0 